+++ common/db/src/main/java/is/codion/common/db/database/Database.java
@@ -86,7 +86,7 @@ public interface Database extends ConnectionFactory {
 	/**
 	 * A comma separated list of paths to scripts to run when initializing the database, implementation specific
 	 */
-	PropertyValue<String> DATABASE_INIT_SCRIPTS = stringValue("codion.db.initScripts");
+	PropertyValue<String> INIT_SCRIPTS = stringValue("codion.db.initScripts");
 
 	/**
 	 * Specifies the timeout (in seconds) to use when checking if database connections are valid.