+++ demos/chinook/src/main/java/is/codion/demos/chinook/tutorial/ClientTutorial.java
@@ -192,7 +192,7 @@ public final class ClientTutorial {
 
 	public static void main(String[] args) {
 		Database.DATABASE_URL.set("jdbc:h2:mem:h2db");
-		Database.DATABASE_INIT_SCRIPTS.set("src/main/sql/create_schema.sql");
+		Database.INIT_SCRIPTS.set("src/main/sql/create_schema.sql");
 		EntityPanel.Config.TOOLBAR_CONTROLS.set(true);
 		FilterTable.AUTO_RESIZE_MODE.set(JTable.AUTO_RESIZE_ALL_COLUMNS);
 		EntityApplicationPanel.builder(ApplicationModel.class, ApplicationPanel.class)