+++ demos/chinook/src/main/java/is/codion/demos/chinook/tutorial/EntitiesTutorial.java
@@ -238,7 +238,7 @@ public final class EntitiesTutorial {
 	public static void main(String[] args) {
 		// Configure the database
 		Database.DATABASE_URL.set("jdbc:h2:mem:h2db");
-		Database.DATABASE_INIT_SCRIPTS.set("src/main/sql/create_schema.sql");
+		Database.INIT_SCRIPTS.set("src/main/sql/create_schema.sql");
 		// initialize a connection provider, this class is responsible
 		// for supplying a valid connection or throwing an exception
 		// in case a connection can not be established