+++ demos/chinook/src/main/java/is/codion/demos/chinook/tutorial/MultiPanelDemo.java
@@ -43,7 +43,7 @@ public final class MultiPanelDemo {
 
 	public static void main(String[] args) {
 		Database.DATABASE_URL.set("jdbc:h2:mem:h2db");
-		Database.DATABASE_INIT_SCRIPTS.set("src/main/sql/create_schema.sql");
+		Database.INIT_SCRIPTS.set("src/main/sql/create_schema.sql");
 
 		LocalEntityConnectionProvider connectionProvider = LocalEntityConnectionProvider.builder()
 						.domain(new ChinookImpl())